bot/requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.5
pytz==2022.6
redmail==0.4.2
yarl==1.8.1
openpyxl==3.0.10
